Refacto de formula-examples pour enlever immutable
