+++ package.json
@@ -39,7 +39,7 @@
     "d3-array": "^2.0.3",
     "d3-fetch": "^1.1.2",
     "fast-memoize": "^2.5.1",
-    "immutable": "^4.0.0-rc.12",
+    "immer": "^3.1.2",
     "lodash-es": "^4.17.11",
     "nearley": "^2.16.0",
     "preact": "^8.4.2"
