+++ scripts/finance/makeLigneBudgetId.js
@@ -1,0 +1,8 @@
+
+export function makeLigneBudgetId(ligneBudget){
+    return [
+        ligneBudget['CodRD'],
+        ligneBudget['Fonction'],
+        ligneBudget['Nature']
+    ].join(' ');
+}
